fix: Change BlockColourBehaviourResolver to BlockColourMapper since it is no longer needed
