--- Assets/Scripts/Blocks/BlockBehaviours/Interfaces/IBlockBehaviourResolver.cs
@@ -1,11 +1,0 @@
-using System;
-using System.Collections.Generic;
-using Assets.Scripts.Blocks;
-
-namespace Assets.Scripts.Blocks
-{
-    public interface IBlockBehaviourResolver
-    {
-        List<Type> Resolve(BlockColour color);
-    }
-}
