--- Assets/Scripts/Blocks/BlockBehaviours/Interfaces/IBlockBehaviourResolver.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4bb40fe7c958d844ca849e737166bdc2
