+++ Assets/Scripts/Blocks/BlockFactory.cs
@@ -7,20 +7,10 @@ namespace Assets.Scripts.Blocks
     public class BlockFactory : MonoBehaviour, IBlockFactory
     {
         [SerializeField] private GameObject _blockPrefab;
-        [SerializeField] private IBlockBehaviourResolver _resolver;
-
-        /// <summary>
-        /// Awake is called when the script instance is being loaded.
-        /// </summary>
-        void Awake()
-        {
-            _resolver = SimpleServiceLocator.Resolve<IBlockBehaviourResolver>();
-        }
 
         public Block SpawnBlock(BlockData data, Transform parent)
         {
             if (_blockPrefab == null) throw new Exception("Block prefab not assigned.");
-            if (_resolver == null) throw new Exception("BlockFactory missing behaviour resolver.");
 
             GameObject go = Instantiate(
                 _blockPrefab,
@@ -32,6 +22,7 @@ namespace Assets.Scripts.Blocks
             Block block = new BlockBuilder(go)
                 .AddBehaviours(data.Behaviours)
                 .WithData(data)
+                .WithColour(BlockColourMapper.ToColour(data.Colour))
                 .Build();
