+++ Assets/Scripts/GameBootstrapper.cs
@@ -9,7 +9,6 @@ using Assets.Scripts.Level;
 public class GameBootstrapper : MonoBehaviour
 {
     private BlockFactory _blockFactory;
-    private IBlockBehaviourResolver _blockBehaviourResolver;
     private BlockWinConditionCounter _blockCounter;
     private ISceneLoader _sceneLoader;
     private LevelDesigner _levelDesigner;
@@ -18,7 +17,6 @@ public class GameBootstrapper : MonoBehaviour
     void Awake()
     {
         _blockFactory = GetComponent<BlockFactory>();
-        _blockBehaviourResolver = new BlockColourBehaviourResolver();
         _blockCounter = new BlockWinConditionCounter();
         _sceneLoader = GetComponent<SceneLoader>();
         _levelDesigner = GetComponent<LevelDesigner>();
@@ -39,7 +37,6 @@ public class GameBootstrapper : MonoBehaviour
     private void RegisterServices()
     {
         SimpleServiceLocator.Register<IBlockFactory>(_blockFactory);
-        SimpleServiceLocator.Register<IBlockBehaviourResolver>(_blockBehaviourResolver);
         SimpleServiceLocator.Register<IBlockCounter>(_blockCounter);
         SimpleServiceLocator.Register<IGameWinCondition>(_blockCounter);
         SimpleServiceLocator.Register<ISceneLoader>(_sceneLoader);
